Refine and cleanup
