+++ go/paddlecloud/get.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 	"text/tabwriter"
 
-	"github.com/PaddlePaddle/cloud/go/utils"
+	"github.com/PaddlePaddle/cloud/go/utils/restclient"
 	"github.com/google/subcommands"
 )
 
@@ -66,7 +66,7 @@ func (p *GetCommand) Execute(_ context.Context, f *flag.FlagSet, _ ...interface{
 func workers(jobname string) error {
 	var queryMap url.Values
 	queryMap.Add("jobname", jobname)
-	respBody, err := utils.GetCall(utils.Config.ActiveConfig.Endpoint+"/api/v1/workers/", queryMap)
+	respBody, err := restclient.GetCall(Config.ActiveConfig.Endpoint+"/api/v1/workers/", queryMap)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error getting workers: %v\n", err)
 		return err
@@ -123,7 +123,7 @@ func registry() error {
 	return err
 }
 func jobs() error {
-	respBody, err := utils.GetCall(utils.Config.ActiveConfig.Endpoint+"/api/v1/jobs/", nil)
+	respBody, err := restclient.GetCall(Config.ActiveConfig.Endpoint+"/api/v1/jobs/", nil)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error getting jobs: %v\n", err)
 		return err
@@ -157,7 +157,7 @@ func jobs() error {
 }
 
 func quota() error {
-	respBody, err := utils.GetCall(utils.Config.ActiveConfig.Endpoint+"/api/v1/quota/", nil)
+	respBody, err := restclient.GetCall(Config.ActiveConfig.Endpoint+"/api/v1/quota/", nil)
 	if err != nil {
 		fmt.Fprintf(os.Stderr, "error getting quota: %v\n", err)
 		return err
