+++ go/paddlecloud/simplefile.go
@@ -11,7 +11,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/PaddlePaddle/cloud/go/utils"
+	"github.com/PaddlePaddle/cloud/go/utils/restclient"
 	"github.com/google/subcommands"
 )
 
@@ -73,7 +73,7 @@ func putFile(src string, dest string) error {
 		destFullPath = dest
 	}
 	query.Set("path", destFullPath)
-	respStr, err := utils.PostFile(utils.Config.ActiveConfig.Endpoint+"/api/v1/file/", src, query)
+	respStr, err := restclient.PostFile(Config.ActiveConfig.Endpoint+"/api/v1/file/", src, query)
 	if err != nil {
 		return err
 	}
@@ -92,16 +92,16 @@ func putFile(src string, dest string) error {
 func getFile(src string, dest string) error {
 	query := url.Values{}
 	query.Set("path", src)
-	req, err := utils.MakeRequestToken(utils.Config.ActiveConfig.Endpoint+"/api/v1/file/", "GET", nil, "", query)
+	req, err := restclient.MakeRequestToken(Config.ActiveConfig.Endpoint+"/api/v1/file/", "GET", nil, "", query)
 	if err != nil {
 		return err
 	}
-	resp, err := utils.HTTPClient.Do(req)
+	resp, err := restclient.HTTPClient.Do(req)
 	if err != nil {
 		return err
 	}
 	defer resp.Body.Close()
-	if resp.Status != utils.HTTPOK {
+	if resp.Status != restclient.HTTPOK {
 		return errors.New("server error: " + resp.Status)
 	}
 	_, srcFile := path.Split(src)
