+++ go/utils/pathutil/pathutil.go
@@ -1,0 +1,18 @@
+package pathutil
+
+import (
+	"os"
+	"runtime"
+)
+
+// UserHomeDir get user home dierctory on different platforms
+func UserHomeDir() string {
+	if runtime.GOOS == "windows" {
+		home := os.Getenv("HOMEDRIVE") + os.Getenv("HOMEPATH")
+		if home == "" {
+			home = os.Getenv("USERPROFILE")
+		}
+		return home
+	}
+	return os.Getenv("HOME")
+}
